chore(i18n): convert app hooks to ts
