--- packages/plugins/i18n/admin/src/contentManagerHooks/addLocaleToCollectionTypesLinks.ts
@@ -1,23 +1,0 @@
-import addLocaleToLinksSearch from './utils/addLocaleToLinksSearch';
-
-const addLocaleToCollectionTypesLinksHook = ({ ctLinks, models }: any, store: any) => {
-  if (!ctLinks.length) {
-    return { ctLinks, models };
-  }
-
-  const storeState = store.getState();
-  const { locales } = storeState.i18n_locales;
-  const { collectionTypesRelatedPermissions } = storeState.rbacProvider;
-
-  const mutatedLinks = addLocaleToLinksSearch(
-    ctLinks,
-    'collectionType',
-    models,
-    locales,
-    collectionTypesRelatedPermissions
-  );
-
-  return { ctLinks: mutatedLinks, models };
-};
-
-export default addLocaleToCollectionTypesLinksHook;
